Move overlay and logo tests off the live source sets
Compose-only overlay cases live in commonTest behind
supportsComposeRuntimeTests. FileUrlTest uses a temporary Path and no
longer loads the native library.
