+++ lib/maplibre-compose/src/commonTest/kotlin/org/maplibre/compose/overlay/MaplibreLogoTest.kt
@@ -4,15 +4,19 @@ import androidx.compose.ui.test.ExperimentalTestApi
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.v2.runComposeUiTest
 import kotlin.test.Test
+import org.maplibre.compose.testing.supportsComposeRuntimeTests
 
 @OptIn(ExperimentalTestApi::class)
 class MaplibreLogoTest {
   // The logo is an Android vector XML, which every target but Android parses with Compose
   // Multiplatform's own reader. Composing it is the only check that the artwork survives that.
   @Test
-  fun logoComposes() = runComposeUiTest {
-    setContent { MaplibreLogo(contentDescription = "MapLibre") }
-    waitForIdle()
-    onNodeWithContentDescription("MapLibre").assertExists()
+  fun logoComposes() {
+    if (!supportsComposeRuntimeTests) return
+    runComposeUiTest {
+      setContent { MaplibreLogo(contentDescription = "MapLibre") }
+      waitForIdle()
+      onNodeWithContentDescription("MapLibre").assertExists()
+    }
   }
 }
